% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_das}
\alias{all_das}
\title{Data availability information of PMC documents}
\format{
A data frame with 94114 rows and 5 variables:
\describe{
\item{month}{month of publication}
\item{year}{year of publication}
\item{PMCID}{unique 7-digit ID associated with document}
\item{das}{a nested tibble with data sharing information}
\item{has_das}{logical value for a document having any data sharing information}
...
}
}
\usage{
all_das
}
\description{
A dataset containing the PMCID of documents published from
January 2017 to December 2019, with data sharing information nested
as tibbles. The data sharing information is either a tibble with
tagging method and content, or a string saying "No data availability
information."
}
\keyword{datasets}
