% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_labeled_das}
\alias{all_labeled_das}
\title{Data availability labels of PMC documents}
\format{
A data frame with 94114 rows and 4 variables:
\describe{
\item{month}{month of publication}
\item{year}{year of publication}
\item{PMCID}{unique 7-digit ID associated with document}
\item{label}{an integer from 0 to 8 describing the data sharing}
...
}
}
\usage{
all_labeled_das
}
\description{
A dataset containing the PMCID of documents published from
January 2017 to December 2019, with a integer label between
0 and 8 that categorizes the data sharing of a document into
one of the 9 groups. The groups are as follows:
}
\details{
\enumerate{
\item No data sharing information.
\item Supplementary materials for download.
\item Data citation or link to repository.
\item Data upon request.
\item Restricted access.
\item Supplementary materials, but no downloadable file.
\item Other.
\item Inline figures or tables.
\item Two or more categories.
}

The groups were generated by passing the \code{all_das} dataset into
the label_DAS function in the package.
}
\keyword{datasets}
