% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_text.R
\name{annotate_text}
\alias{annotate_text}
\title{Use text mining and regular expressions to extract data sharing information from a document}
\usage{
annotate_text(pmcid)
}
\arguments{
\item{pmcid}{A string of the PMCID number of the document}
}
\value{
A tibble with the tokenized sentences and logical columns is_available,
not_available, external_source, supplementary, and has_das that are TRUE or FALSE
for each sentence.
}
\description{
This function uses text mining and regular expressions to determine, sentence
by sentence, if there is a match with specific data availability keywords.
The algorithm used in this function is inspired by the ODDPub package, which
is available on GitHub at github.com/quest-bih/oddpub. There are vectors of
keywords which are grouped as words associated with: available, was_available,
not_available, repository, data, not_data, source_code, supplements_and_files,
and data_availability. An input document is tokenized by sentences and each
sentence is marked if it matches one of the five following combinations:
}
\details{
\enumerate{
\item is_available = data NEAR (available OR was_available)
\item not_available = not_data OR not_available
\item external_source = repository OR source_code
\item supplementary = supplements_and_files
\item has_das = data_availability
}

x NEAR y means that keyword x is detected in the same sentence, the following sentence,
or the previous sentence as keyword y.
}
\examples{
annotate_text("PMC7437466")

}
