% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_labels.R
\name{balance_labels}
\alias{balance_labels}
\title{Balance a dataset using the ROSE algorithm}
\usage{
balance_labels(das_df, p)
}
\arguments{
\item{das_df}{A full data matrix, with labels and input features.}

\item{p}{The probability of the minority class examples in the
resulting data set generated by ROSE. p should be between 0 and 1.}
}
\value{
A resampled dataset, with balanced labels.
}
\description{
The data sharing labels in PMC from 2018 are heavily imbalanced.
Most of the documents either have no data sharing information or
supplementary material (labels 0 and 1 from the \code{label_DAS} function).
This function applies the Random Over-Sampling Experiment (ROSE) algorithm
to generate synethically oversampled datapoints of groups 2 through 6.
}
\examples{
balanced_data <- balance_labels(my_data, p = 0.6)

}
