% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_DAS.R
\name{clean_DAS}
\alias{clean_DAS}
\title{Clean DAS information created by the \code{get_DAS} function}
\usage{
clean_DAS(das_df)
}
\arguments{
\item{das_df}{A tibble generated by \code{get_DAS}}
}
\value{
A tibble with the columns PMCID, tagging_method, content,
has_das, and attached_media. This resulting tibble is ready to be passed
into the function \code{label_DAS}.
}
\description{
This function cleans and preps data sharing information to use
to later label the PMCIDs by data sharing type.
}
\examples{
get_DAS("PMC3257301") \%>\% 
clean_DAS()

}
