% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_id.R
\name{convert_id}
\alias{convert_id}
\title{Convert between PMCID (PubMed Central) and PMID (PubMed)}
\usage{
convert_id(id, from, to)
}
\arguments{
\item{id}{A string of ID number}

\item{from}{The database that the ID number is from}

\item{to}{The database we want to the convert the ID number to}
}
\value{
A string of the ID number for the correct database
}
\description{
This function utilizes the ID converter provided by the NCBI
to switch between a PMCID to PMID or PMID to PMCID.
}
\examples{
convert_id("3256301", from = "pmc", to = "pubmed")
convert_id("3256301", from = "pmcid", to = "pmid")
convert_id("22253597", from = "pubmed", to = "pmc")
convert_id("22253597", from = "pmcid", to = "pmid")

}
