% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{das_words}
\alias{das_words}
\title{All words contained in the data availability section of a document}
\format{
A data frame with 19930498 rows and 2 variables:
\describe{
\item{PMCID}{unique 7-digit ID associated with document}
\item{word}{an English word that appears in the document's data section}
...
}
}
\usage{
das_words
}
\description{
A dataset containing all the words in data availability section of
documents published between January 2017 and December 2019. During the tagging
process, sections of documents are pulled out that are related to data sharing,
whether it be an actual Data Availability Statement or an inline figure.
The text content of these sections is made available in this dataset.
}
\keyword{datasets}
