% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_features.R
\name{generate_features}
\alias{generate_features}
\title{Generate numeric features from text data to use in modeling}
\usage{
generate_features(df, lem_stem, mode)
}
\arguments{
\item{df}{A tibble with two columns: PMCID and text generated by
\code{tidypmc::pmc_text}.}

\item{lem_stem}{A string that is one of "lem" or "stem" which tells
the function to respectively lemmatize or stem the words}

\item{mode}{A string that is one of: 'bow', 'tfidf', or 'tm n' where
n is an integer number. The 'bow' mode creates a document-term matrix
with counts. The 'tfidf' mode creates a document-term matrix with
term frequency inverse document frequency values. The 'tm n' mode
creates a matrix with topic modeling embeddings, generated by an
unsupervised Latent Dirichlet allocation algorithm. n is an integer value
that specifies the number of topics that we use in the LDA model. In this
mode, each row is a document and each column is a topic model feature.}
}
\value{
A document-term matrix if mode is "bow" or "tfidf" or a matrix with
LDA topic model features if mode is "tm n".
}
\description{
This function generates features depending on the input mode, to
embed text data as numeric features. This function also serves as
a general preprocessing step that tokenizes the text, removes
stopwords, and either lemmatizes or stems the words (depending on input).
}
\examples{
input_df <-
tribble(
   ~PMCID, ~xml,
   "PMC3257301", xml_by_id("PMC3257301")
) \%>\% 
mutate(
   text = pmc_text(xml)
) \%>\% 
select(PMCID, text)

generate_features(input_df, lem_stem = "lem", mode = "tm 10")
generate_features(input_df, lem_stem = "lem", mode = "tfidf")
generate_features(input_df, lem_stem = "stem", mode = "bow")

}
