% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abstract.R
\name{get_abstract}
\alias{get_abstract}
\title{Get abstract of a paper by PMCID}
\usage{
get_abstract(id)
}
\arguments{
\item{id}{A string of PMCID number of the document}
}
\value{
A string that is the full abstract text.
}
\description{
This is just a wrapper function for \code{tidypmc::pmc_text} by Chris Stubben
to get just the abstract text of a document.
}
\examples{
get_abstract("PMC4161373")

}
