% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DAS.R
\name{get_DAS}
\alias{get_DAS}
\title{Get data sharing and accessibility information of a document}
\usage{
get_DAS(id)
}
\arguments{
\item{id}{A string of the PMCID number of the document}
}
\value{
A tibble with two columns: tagging method (matching one of the three
methods above) and content. If there is no data sharing information in the document
that matches the methods above, return the string "No data availability information".
}
\description{
This function uses the PMCID to get the full-text PMC XML document. It then
uses text mining, regular expressions, and XML tagging to return data sharing
information. The possible ways that this function detects data sharing information:
(1) A specific Data Availability Statement section is tagged in the XML,
(2) A Supplementary Materials section is tagged in the XML,
(3) Any section of the document has a title matching a list of titles included below.
}
\examples{
get_DAS("PMC4161373")

}
