% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funding.R
\name{get_funding}
\alias{get_funding}
\title{Get grant information of a paper by PMCID}
\usage{
get_funding(id)
}
\arguments{
\item{id}{A string of the PMCID number of the document}
}
\value{
A tibble with one column containing the grant agency.
}
\description{
This function uses the ID converstion API provided by the NCBI
to get the PubMed document associated with a PubMed Central document.
It then searches the full PubMed XML document for the tagged Grants
section and returns the grant information.
}
\examples{
get_funding("PMC4161373")

}
