% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mesh.R
\name{get_mesh}
\alias{get_mesh}
\title{Get Medical Subject Heading (MeSH) terms of a document by PMCID}
\usage{
get_mesh(id)
}
\arguments{
\item{id}{A string of the PMCID number of the document}
}
\value{
A tibble with one column containing the associated MeSH terms.
}
\description{
This function uses the ID converstion API provided by the NCBI
to get the PubMed document associated with a PubMed Central document.
It then searches the full PubMed XML document for tagged MeSH terms and returns
the MeSH terms associated with the document.
}
\examples{
get_mesh("PMC4161373")

}
