% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_references.R
\name{get_references}
\alias{get_references}
\title{Get references by paper PMCID}
\usage{
get_references(id)
}
\arguments{
\item{id}{A string of PMCID number of the document}
}
\value{
A tibble with id, pmid, authors, year, title, journal,
volume pages, and doi.
}
\description{
This is just a wrapper function for \code{tidypmc::pmc_reference()} by Chris Stubben
to get all reference information from a PMCID.
}
\examples{
get_references("PMC4161373")

}
