% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_DAS.R
\name{label_DAS}
\alias{label_DAS}
\title{Labels documents by data sharing class}
\usage{
label_DAS(das_df)
}
\arguments{
\item{das_df}{A tibble generated by \code{get_DAS} and \code{clean_DAS}}
}
\value{
A tibble with the columns PMCID and label
}
\description{
\enumerate{
\item No data sharing information.
\item Supplementary materials for download.
\item Data citation or link to repository.
\item Data upon request.
\item Restricted access.
\item Supplementary materials, but no downloadable file.
\item Other.
\item Inline figures or tables.
\item Two or more categories.
}
}
\examples{
get_DAS("PMC3257301") \%>\% 
clean_DAS() \%>\% 
label_DAS()

}
