% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_features.R
\name{mesh_features}
\alias{mesh_features}
\title{Create binary encoding of a paper's MeSH terms}
\usage{
mesh_features(id)
}
\arguments{
\item{id}{A string of the PMCID number of the document}
}
\value{
A 1 x 111 matrix that has 1's corresponding to the included MeSH terms.
}
\description{
This function uses the assigns a numerical embedding of MeSH
(Medical Subject Heading) terms. MeSH terms form a tree structure,
where some tags are sub-categories of a more broad tag. MeSH tags range from
high-level terms like "Metabolism (G03)" to "DNA Methylation (G03.059.538.161)".
This function focuses on the 111 high level terms, from "Body Regions (A01)"
to "Geographic Locations (Z01)". This function takes all the more specific tags
of a document (represented by ID number) and matches them to their parent/higher
level term. It then uses binary encoding to numerically embed the MeSH terms,
where it will have a 1 if the document contains the MeSH term and 0 otherwise.
}
\examples{
mesh_features("PMC4161373")

}
