% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mesh_table}
\alias{mesh_table}
\title{MeSH terms by number, with paths in the tree structure}
\format{
A data frame with 60666 rows and 10 variables:
\describe{
\item{node1}{first level of the tree, a letter between A and Z, followed by a numerical value}
\item{node2}{second level of the tree, a three digit number. For nodes 2 through 8, NA if further down the tree than the MeSH term.}
\item{node3}{third level of the tree, a three digit number}
\item{node4}{fourth level of the tree, a three digit number}
\item{node5}{fifth level of the tree, a three digit number}
\item{node6}{sixth level of the tree, a three digit number}
\item{node7}{seventh level of the tree, a three digit number}
\item{node8}{eighth level of the tree, a three digit number}
\item{description}{the MeSH term}
\item{pathString}{a path of nodes 1 through 8}
...
}
}
\usage{
mesh_table
}
\description{
A dataset containing all the MeSH terms and their associated numbers.
MeSH terms are numbered using a tree, and this dataset provides the
different levels of the tree associated with a term. This dataset was
created from \url{https://www.nlm.nih.gov/databases/download/mesh.html}.
}
\keyword{datasets}
