% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mesh_top_level}
\alias{mesh_top_level}
\title{The highest level MeSH terms with associated the first level of the MeSH tree}
\format{
A data frame with 111 rows and 2 variables:
\describe{
\item{descriptor}{the MeSH term}
\item{branch}{the branch number containing a letter A through Z, followed by a two digit number}
...
}
}
\usage{
mesh_top_level
}
\description{
A dataset containing only the highest level MeSH terms, with the
associated branch number.
}
\keyword{datasets}
