% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_text.R
\name{preprocess_text}
\alias{preprocess_text}
\title{Normalize and preprocess numeric features}
\usage{
preprocess_text(mat, mode, type)
}
\arguments{
\item{mat}{A matrix (either document-term matrix or topic model matrix)
generated by \code{generate_features}}

\item{mode}{A string that either has the value 'pca n' or 'none' where
n is an integer number specifying the number of principle components
from the principle component decomposition.}

\item{type}{A string that is either 'dtm' or 'none'. type should be 'dtm'
if we generated a document-term matrix in \code{generate_features} (from bow
or tfidf) and 'none' if we generated a topic model matrix.}
}
\value{
A normalized and centered feature matrix, with possible dimensionality
reduction if mode = 'pca n'. If mode = 'pca n', the function will also print the
principle component information.
}
\description{
This function normalizes and centers the input matrix (by rows), after we've
generated features via the \code{generate_features} function. There's also
the option of performing dimensionality reduction by Principle Component
Analysis.
}
\examples{
input_mat <-
tribble(
   ~PMCID, ~xml,
   "PMC3257301", xml_by_id("PMC3257301")
) \%>\% 
mutate(
   text = pmc_text(xml)
) \%>\% 
select(PMCID, text) \%>\% 
generate_features(input_df, lem_stem = "lem", mode = "tfidf")

preprocess_text(input_mat, mode = "none", type = "dtm")
preprocess_text(input_mat, mode = "pca 50", type = "dtm")

}
