% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_by_date.R
\name{search_by_date}
\alias{search_by_date}
\title{Search a database by date}
\usage{
search_by_date(start, end)
}
\arguments{
\item{start}{A string specifying the start date. The string must be formatted as
YYYY, YYY/MM, or YYYY/MM/DD.}

\item{end}{A string specifying the end date. The string must be formatted as
YYYY, YYY/MM, or YYYY/MM/DD.}
}
\value{
A tibble with one column (labeled PMCID) that contains all the PMCID
numbers returned from the database search.
}
\description{
Search PubMed Central by a start and end date to get a list of documents
that were published in PMC within the date window. Uses the ESearch
functionality of E-utilities to search a specified database. This function
returns a maximum of 100,000 documents that fall within the window.
}
\examples{
search_by_date("2020/01/01", "2020/01/31")
search_by_date("2020/01/01", "2020/01/01")

}
