% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_by_mesh.R
\name{search_by_mesh}
\alias{search_by_mesh}
\title{Search a database by a MeSH term}
\usage{
search_by_mesh(mesh_term, n = 20)
}
\arguments{
\item{n}{An integer number of document IDs to return, defaults to 20.}

\item{keyword}{A string of a MeSH term.}
}
\value{
A tibble with one column (labeled PMCID) that contains all the PMCID
numbers returned from the database search.
}
\description{
Search a database by a MeSH term to get a list of documents
that correspond to the input keyword. Uses the ESearch functionality of
E-utilities to search PubMed Central by MeSH.
}
\examples{
search_by_mesh("Oceania")
search_by_mesh("Multiple Birth Offspring")

}
