% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_metadata.R
\name{tidy_metadata}
\alias{tidy_metadata}
\title{Get a document's metadata in a tidy format}
\usage{
tidy_metadata(id)
}
\arguments{
\item{id}{A string specifying the ID number corresponding to a document.}
}
\value{
a tibble with two columns: the metadata field and the value.
}
\description{
Wrapper function for \code{tidypmc::pmc_metadata()} to return metadata in a tidy format.
}
\examples{
test_id <- "PMC7344926"
tidy_metadata(test_id)

}
