% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_by_id.R
\name{xml_by_id}
\alias{xml_by_id}
\title{Download full text XML from PubMed Central by ID number}
\usage{
xml_by_id(id, db = "pmc")
}
\arguments{
\item{id}{A string specifying the ID number corresponding to a document.}

\item{db}{A string specifying the dataset to pull the XML from, defaults to "pmc" (PubMed Central).
Other database options include "pubmed" (PubMed), "mesh" (MeSH), and "nlmcatalog" (NLM Catalog).
The full list of database options are available on the NCBI E-utilities guide.}
}
\value{
an XML document nodeset (as defined by the \code{xml2} package).
}
\description{
Fetches a full text XML specified by ID number from the E-utilities API.
}
\examples{
full_text_xml <- xml_by_id("PMC7344926")

}
